scrits/api/base/serializers.py
# -*- coding: utf-8 -*-
from __future__ import absolute_import

from django.contrib.auth import get_user_model

from rest_framework import serializers


class AuthorSerializer(serializers.ModelSerializer):

    class Meta:
        model = get_user_model()
        fields = (
            'id',
            'first_name',
            'last_name',
        )
        read_only_fields = ('first_name', 'last_name')
